WIP sse 3

+++ imgcoder.h
@@ -10,10 +10,10 @@ typedef struct
 {
     unsigned char mb_x;
     unsigned char mb_y;
-    double rms;
+    float rms;
     unsigned char blockData[MB_SIZE][MB_SIZE][3];
-    double blockDataDCT[MB_SIZE][MB_SIZE][4] __attribute__((aligned(16)));  // Red Green Blue Unused
-    double* rleData[3];
+    float blockDataDCT[MB_SIZE][MB_SIZE][4] __attribute__((aligned(16)));  // Red Green Blue Unused
+    float* rleData[3];
     int rleSize[3];
 } macroblock_t;
 
@@ -22,8 +22,8 @@ typedef struct
     unsigned char mb_x;
     unsigned char mb_y;
     int rleSize[3];
-    double* rleData[3];
-    double blockDataDCT[MB_SIZE][MB_SIZE][4] __attribute__((aligned(16)));  // Red Green Blue Unused
+    float* rleData[3];
+    float blockDataDCT[MB_SIZE][MB_SIZE][4] __attribute__((aligned(16)));  // Red Green Blue Unused
 } compressed_macroblock_t;
 
 compressed_macroblock_t* encodeImage(const unsigned char* imgIn, const unsigned char* prevFrame, macroblock_t* blocks, unsigned char* rmsView);
